Rename widgets routes info from template to routes
